add new file fsntypes.go to put fsn types separately
