+++ common/types.go
@@ -340,365 +340,3 @@ func (ma *MixedcaseAddress) ValidChecksum() bool {
 func (ma *MixedcaseAddress) Original() string {
 	return ma.original
 }
-
-// FSNCallAddress wacom
-var FSNCallAddress = HexToAddress("0xffffffffffffffffffffffffffffffffffffffff")
-
-// TicketLogAddress wacom
-var TicketLogAddress = HexToAddress("0xfffffffffffffffffffffffffffffffffffffffe")
-
-// SystemAssetID wacom
-var SystemAssetID = HexToHash("0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff")
-
-// OwnerUSANAssetID wacom
-var OwnerUSANAssetID = HexToHash("0xfffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe")
-
-// NotationKeyAddress wacom
-var NotationKeyAddress = HexToAddress("0xfffffffffffffffffffffffffffffffffffffffd")
-
-// AssetKeyAddress wacom
-var AssetKeyAddress = HexToAddress("0xfffffffffffffffffffffffffffffffffffffffc")
-
-// TicketKeyAddress wacom
-var TicketKeyAddress = HexToAddress("0xfffffffffffffffffffffffffffffffffffffffb")
-
-// SwapKeyAddress wacom
-var SwapKeyAddress = HexToAddress("0xfffffffffffffffffffffffffffffffffffffffa")
-
-// MultiSwapKeyAddress wacom
-var MultiSwapKeyAddress = HexToAddress("0xfffffffffffffffffffffffffffffffffffffff9")
-
-// ReportIllegalAddress wacom
-var ReportKeyAddress = HexToAddress("0xfffffffffffffffffffffffffffffffffffffff8")
-
-func (addr Address) IsSpecialKeyAddress() bool {
-	return addr == TicketKeyAddress ||
-		addr == NotationKeyAddress ||
-		addr == AssetKeyAddress ||
-		addr == SwapKeyAddress ||
-		addr == MultiSwapKeyAddress ||
-		addr == ReportKeyAddress
-}
-
-var (
-	// AutoBuyTicket wacom
-	AutoBuyTicket = false
-	// AutoBuyTicketChan wacom
-	AutoBuyTicketChan = make(chan int, 10)
-
-	// ReportIllegal wacom
-	ReportIllegalChan = make(chan []byte)
-)
-
-// FSNCallFunc wacom
-type FSNCallFunc uint8
-
-const (
-	// GenNotationFunc wacom
-	GenNotationFunc = iota
-	// GenAssetFunc wacom
-	GenAssetFunc
-	// SendAssetFunc wacom
-	SendAssetFunc
-	// TimeLockFunc wacom
-	TimeLockFunc
-	// BuyTicketFunc wacom
-	BuyTicketFunc
-	// OldAssetValueChangeFunc wacom (deprecated)
-	OldAssetValueChangeFunc
-	// MakeSwapFunc wacom
-	MakeSwapFunc
-	// RecallSwapFunc wacom
-	RecallSwapFunc
-	// TakeSwapFunc wacom
-	TakeSwapFunc
-	// EmptyFunc wacom
-	EmptyFunc
-	// MakeSwapFuncExt wacom
-	MakeSwapFuncExt
-	// TakeSwapFuncExt wacom
-	TakeSwapFuncExt
-	// AssetValueChangeFunc wacom
-	AssetValueChangeFunc
-	// MakeMultiSwapFunc wacom
-	MakeMultiSwapFunc
-	// RecallMultiSwapFunc wacom
-	RecallMultiSwapFunc
-	// TakeMultiSwapFunc wacom
-	TakeMultiSwapFunc
-	// ReportIllegalFunc wacom
-	ReportIllegalFunc
-)
-
-func (f FSNCallFunc) Name() string {
-	switch f {
-	case GenNotationFunc:
-		return "GenNotationFunc"
-	case GenAssetFunc:
-		return "GenAssetFunc"
-	case SendAssetFunc:
-		return "SendAssetFunc"
-	case TimeLockFunc:
-		return "TimeLockFunc"
-	case BuyTicketFunc:
-		return "BuyTicketFunc"
-	case OldAssetValueChangeFunc:
-		return "OldAssetValueChangeFunc"
-	case MakeSwapFunc:
-		return "MakeSwapFunc"
-	case RecallSwapFunc:
-		return "RecallSwapFunc"
-	case TakeSwapFunc:
-		return "TakeSwapFunc"
-	case EmptyFunc:
-		return "EmptyFunc"
-	case MakeSwapFuncExt:
-		return "MakeSwapFuncExt"
-	case TakeSwapFuncExt:
-		return "TakeSwapFuncExt"
-	case AssetValueChangeFunc:
-		return "AssetValueChangeFunc"
-	case MakeMultiSwapFunc:
-		return "MakeMultiSwapFunc"
-	case RecallMultiSwapFunc:
-		return "RecallMultiSwapFunc"
-	case TakeMultiSwapFunc:
-		return "TakeMultiSwapFunc"
-	case ReportIllegalFunc:
-		return "ReportIllegalFunc"
-	}
-	return "Unkonwn"
-}
-
-func IsFsnCall(to *Address) bool {
-	return to != nil && *to == FSNCallAddress
-}
-
-func GetFsnCallFee(to *Address, funcType FSNCallFunc) *big.Int {
-	fee := big.NewInt(0)
-	if !IsFsnCall(to) {
-		return fee
-	}
-	switch funcType {
-	case GenNotationFunc:
-		fee = big.NewInt(100000000000000000) // 0.1 FSN
-	case GenAssetFunc:
-		fee = big.NewInt(10000000000000000) // 0.01 FSN
-	case MakeSwapFunc, MakeSwapFuncExt, MakeMultiSwapFunc:
-		fee = big.NewInt(1000000000000000) // 0.001 FSN
-	case TimeLockFunc:
-		fee = big.NewInt(1000000000000000) // 0.001 FSN
-	}
-	return fee
-}
-
-// ParseBig256 parses s as a 256 bit integer in decimal or hexadecimal syntax.
-// Leading zeros are accepted. The empty string parses as zero.
-func ParseBig256(s string) (*big.Int, bool) {
-	if s == "" {
-		return new(big.Int), true
-	}
-	var bigint *big.Int
-	var ok bool
-	if len(s) >= 2 && (s[:2] == "0x" || s[:2] == "0X") {
-		bigint, ok = new(big.Int).SetString(s[2:], 16)
-	} else {
-		bigint, ok = new(big.Int).SetString(s, 10)
-	}
-	if ok && bigint.BitLen() > 256 {
-		bigint, ok = nil, false
-	}
-	return bigint, ok
-}
-
-// ToAsset wacom
-func (p *GenAssetParam) ToAsset() Asset {
-	return Asset{
-		Name:        p.Name,
-		Symbol:      p.Symbol,
-		Decimals:    p.Decimals,
-		Total:       p.Total,
-		CanChange:   p.CanChange,
-		Description: p.Description,
-	}
-}
-
-// Asset wacom
-type Asset struct {
-	ID          Hash
-	Owner       Address
-	Name        string
-	Symbol      string
-	Decimals    uint8
-	Total       *big.Int `json:",string"`
-	CanChange   bool
-	Description string
-}
-
-func (u *Asset) DeepCopy() Asset {
-	total := *u.Total
-	return Asset{
-		ID:        u.ID,
-		Owner:     u.Owner,
-		Name:      u.Name,
-		Symbol:    u.Symbol,
-		Decimals:  u.Decimals,
-		Total:     &total,
-		CanChange: u.CanChange,
-	}
-}
-
-func (u *Asset) MarshalJSON() ([]byte, error) {
-	return json.Marshal(&struct {
-		ID          Hash
-		Owner       Address
-		Name        string
-		Symbol      string
-		Decimals    uint8
-		Total       string
-		CanChange   bool
-		Description string
-	}{
-		ID:          u.ID,
-		Owner:       u.Owner,
-		Name:        u.Name,
-		Symbol:      u.Symbol,
-		Decimals:    u.Decimals,
-		Total:       u.Total.String(),
-		CanChange:   u.CanChange,
-		Description: u.Description,
-	})
-}
-
-// SystemAsset wacom
-var SystemAsset = Asset{
-	Name:        "Fusion",
-	Symbol:      "FSN",
-	Decimals:    18,
-	Total:       new(big.Int).Mul(big.NewInt(81920000), big.NewInt(1000000000000000000)),
-	ID:          SystemAssetID,
-	Description: "https://fusion.org",
-}
-
-// Swap wacom
-type Swap struct {
-	ID            Hash
-	Owner         Address
-	FromAssetID   Hash
-	FromStartTime uint64
-	FromEndTime   uint64
-	MinFromAmount *big.Int `json:",string"`
-	ToAssetID     Hash
-	ToStartTime   uint64
-	ToEndTime     uint64
-	MinToAmount   *big.Int `json:",string"`
-	SwapSize      *big.Int `json:",string"`
-	Targes        []Address
-	Time          *big.Int // Provides information for TIME
-	Description   string
-	Notation      uint64
-}
-
-// DeepCopy wacom
-func (s *Swap) DeepCopy() Swap {
-	minFromAmount := *s.MinFromAmount
-	minToAmount := *s.MinToAmount
-	swapSize := *s.SwapSize
-	swapTime := *s.Time
-	targets := make([]Address, len(s.Targes))
-	copy(targets, s.Targes)
-
-	return Swap{
-		ID:            s.ID,
-		Owner:         s.Owner,
-		FromAssetID:   s.FromAssetID,
-		FromStartTime: s.FromStartTime,
-		FromEndTime:   s.FromEndTime,
-		MinFromAmount: &minFromAmount,
-		ToAssetID:     s.ToAssetID,
-		ToStartTime:   s.ToStartTime,
-		ToEndTime:     s.ToEndTime,
-		MinToAmount:   &minToAmount,
-		SwapSize:      &swapSize,
-		Targes:        targets,
-		Time:          &swapTime,
-		Description:   s.Description,
-	}
-}
-
-// MultiSwap wacom
-type MultiSwap struct {
-	ID            Hash
-	Owner         Address
-	FromAssetID   []Hash
-	FromStartTime []uint64
-	FromEndTime   []uint64
-	MinFromAmount []*big.Int `json:",string"`
-	ToAssetID     []Hash
-	ToStartTime   []uint64
-	ToEndTime     []uint64
-	MinToAmount   []*big.Int `json:",string"`
-	SwapSize      *big.Int   `json:",string"`
-	Targes        []Address
-	Time          *big.Int // Provides information for TIME
-	Description   string
-	Notation      uint64
-}
-
-// DeepCopy wacom
-func (s *MultiSwap) DeepCopy() MultiSwap {
-	minFromAmount := make([]*big.Int, len(s.MinFromAmount))
-	for k, v := range s.MinFromAmount {
-		minFromAmount[k] = new(big.Int).SetBytes(v.Bytes())
-	}
-	minToAmount := make([]*big.Int, len(s.MinToAmount))
-	for k, v := range s.MinToAmount {
-		minToAmount[k] = new(big.Int).SetBytes(v.Bytes())
-	}
-	swapSize := *s.SwapSize
-	swapTime := *s.Time
-	targets := make([]Address, len(s.Targes))
-	copy(targets, s.Targes)
-
-	return MultiSwap{
-		ID:            s.ID,
-		Owner:         s.Owner,
-		FromAssetID:   s.FromAssetID,
-		FromStartTime: s.FromStartTime,
-		FromEndTime:   s.FromEndTime,
-		MinFromAmount: minFromAmount,
-		ToAssetID:     s.ToAssetID,
-		ToStartTime:   s.ToStartTime,
-		ToEndTime:     s.ToEndTime,
-		MinToAmount:   minToAmount,
-		SwapSize:      &swapSize,
-		Targes:        targets,
-		Time:          &swapTime,
-		Description:   s.Description,
-	}
-}
-
-func CheckSwapTargets(targets []Address, addr Address) error {
-	if len(targets) == 0 {
-		return nil
-	}
-	for _, target := range targets {
-		if addr == target {
-			return nil
-		}
-	}
-	return fmt.Errorf("swap taker does not match the specified targets")
-}
-
-// KeyValue wacom
-type KeyValue struct {
-	Key   string
-	Value interface{}
-}
-
-// NewKeyValue wacom
-func NewKeyValue(name string, v interface{}) *KeyValue {
-
-	return &KeyValue{Key: name, Value: v}
-
-}
